Added OBJ format loader (limited support)
Hacked some colors (need to add light)
Fixed aspect ratio when window size is different from requested

+++ Object3D/Object3D.hpp
@@ -34,7 +34,7 @@ class Object3D
         /**
          * Returns the colors array information
          */
-        virtual const GLubyte *getIndicesArray() = 0;
+        virtual const GLuint *getIndicesArray() = 0;
         virtual uint32_t getIndicesArrayLen() = 0;
 
 		/**
